package refactoring: global
