+++ src/main/java/com/example/week06/global/security/JwtAuthProvider.java
@@ -1,7 +1,7 @@
 package com.example.week06.global.security;
 
-import com.example.week06.model.User;
-import com.example.week06.repository.UserRepository;
+import com.example.week06.domain.user.entity.User;
+import com.example.week06.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.AuthenticationProvider;
